Increase cyclomatic for catch clause and complex lambdas
Increase only on catch clause, as recommended in https://www.ndepend.com/docs/code-metrics#CC
For lambdas, do not increase CC for the lambda, bot only for &&, ||, and, or. The reason is that we don't want to increase CC for statements like Select(x => x.Id). But the current calcullation will not increase CC for lambdas like Where(x => x.Id == 2), which probably should.